Use compiler integrated async phase under -Xasync
